requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-spectacular==0.26.5
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
oauthlib==3.2.2
phonenumbers==8.13.25
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
